Get Spreadsheet files from Box; Remove Box API from requirements; Add celery task for import spreadsheets;
