+++ balance/local_settings.example.py
@@ -1,3 +1,4 @@
 CORS_ORIGIN_WHITELIST = (
     'localhost:4200'
 )
+DEBUG = True
