+++ cards/urls.py
@@ -18,6 +18,5 @@ from django.conf.urls import url
 from cards import views
 
 urlpatterns = [
-    url(r'^retrieve', views.RetrieveBalanceView.as_view()),
-    url(r'^import', views.ImportSpreadsheetView.as_view())
+    url(r'^retrieve', views.RetrieveBalanceView.as_view())
 ]
