+++ requirements.txt
@@ -1,4 +1,3 @@
-boxsdk==1.5.3
 celery==4.0.2
 django==1.10.5
 django-cors-headers==2.0.2
